fix: cast automatico inválido, resposta da associação paciente - profissional de saude

+++ api/src/main/java/com/avicia/api/features/sistema/token/TokenService.java
@@ -31,7 +31,7 @@ public class TokenService {
                 .subject(identity)
                 .expiresAt(generateExpirationDate())
                 .issuedAt(Instant.now())
-                .claim("idUsuario", idUsuario)
+                .claim("idUsuario", idUsuario.intValue())
                 .claim("role", role.getNome())
                 .claim("permissoes", role.getPermissoes())
                 .build();
